Changes introduced to  * do_command.c, cron.h, cron.8: Change the behaviour when logging the    information of the child processes. A new loglevel (8) is introduced and    documented in cron.8 to introduced the process ID of the child    processes in the log. The previous log format is kept unless the sysadmin   choses to select this new option.   (Closes: 637295)  * debian/examples/cron-stats.pl: Change the Perl script to work with the new    behaviour of cron when logging with loglevel
With this change, the startup of the child is logged right  after the fork (by
the parent) to ensure that the logged PID is that of the parent process and the
additional pid is the child's.

Previously, there was no way to differentiate the PID process as the
child would log (after the fork) the same PID twice.